More tests PT 2
main project on mac, i am pushing project to origin, and then on windows computer, i am pulling origin and testing to see if unity opens with the correct changes.  It wasn't working before, but now it is working after I removed project from list in unity hub and added project in from cloned repo. Now this test is to see if being a collaborator or not has any affect on the puller's ability to pull from main.